money type changed to decimal (#71)

+++ source/Manga.Application/Boundaries/GetAccountDetails/Transaction.cs
@@ -4,12 +4,12 @@ namespace Manga.Application.Boundaries.GetAccountDetails
     public sealed class Transaction
     {
         public string Description { get; }
-        public double Amount { get; }
+        public decimal Amount { get; }
         public DateTime TransactionDate { get; }
 
         public Transaction(
             string description,
-            double amount,
+            decimal amount,
             DateTime transactionDate)
         {
             Description = description;
