+++ tests/Manga.UnitTests/UseCaseTests/Withdraw/PositiveDataSetup.cs
@@ -2,7 +2,7 @@ namespace Manga.UnitTests.UseCasesTests.Withdraw
 {
     using Xunit;
 
-    internal sealed class PositiveDataSetup : TheoryData<double, double>
+    internal sealed class PositiveDataSetup : TheoryData<decimal, decimal>
     {
         public PositiveDataSetup()
         {
